Move deadlock_detection and maintain_queue_wait_times to schedulers module
