+++ libs/schedulers/CMakeLists.txt
@@ -9,9 +9,11 @@ cmake_minimum_required(VERSION 3.3.2 FATAL_ERROR)
 list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_SOURCE_DIR}/cmake")
 
 set(schedulers_headers
+  hpx/schedulers/deadlock_detection.hpp
   hpx/schedulers/local_priority_queue_scheduler.hpp
   hpx/schedulers/local_queue_scheduler.hpp
   hpx/schedulers/lockfree_queue_backends.hpp
+  hpx/schedulers/maintain_queue_wait_times.hpp
   hpx/schedulers/queue_helpers.hpp
   hpx/schedulers/shared_priority_queue_scheduler.hpp
   hpx/schedulers/static_priority_queue_scheduler.hpp
@@ -21,9 +23,11 @@ set(schedulers_headers
   hpx/schedulers.hpp)
 
 set(schedulers_compat_headers
+  hpx/runtime/threads/policies/deadlock_detection.hpp
   hpx/runtime/threads/policies/local_priority_queue_scheduler.hpp
   hpx/runtime/threads/policies/local_queue_scheduler.hpp
   hpx/runtime/threads/policies/lockfree_queue_backends.hpp
+  hpx/runtime/threads/policies/maintain_queue_wait_times.hpp
   hpx/runtime/threads/policies/queue_helpers.hpp
   hpx/runtime/threads/policies/shared_priority_queue_scheduler.hpp
   hpx/runtime/threads/policies/static_priority_queue_scheduler.hpp
