+++ libs/schedulers/include/hpx/schedulers/local_queue_scheduler.hpp
@@ -14,15 +14,15 @@
 #include <hpx/affinity/affinity_data.hpp>
 #include <hpx/assertion.hpp>
 #include <hpx/errors.hpp>
+#include <hpx/functional.hpp>
 #include <hpx/logging.hpp>
-#include <hpx/runtime/threads/policies/deadlock_detection.hpp>
+#include <hpx/schedulers/deadlock_detection.hpp>
 #include <hpx/schedulers/lockfree_queue_backends.hpp>
 #include <hpx/schedulers/thread_queue.hpp>
 #include <hpx/threading_base/scheduler_base.hpp>
 #include <hpx/threading_base/thread_data.hpp>
 #include <hpx/threading_base/thread_queue_init_parameters.hpp>
 #include <hpx/topology/topology.hpp>
-#include <hpx/util_fwd.hpp>
 
 #include <atomic>
 #include <cstddef>
