+++ libs/schedulers/include/hpx/schedulers/shared_priority_queue_scheduler.hpp
@@ -16,6 +16,7 @@
 #include <hpx/basic_execution/this_thread.hpp>
 #include <hpx/debugging/print.hpp>
 #include <hpx/errors.hpp>
+#include <hpx/functional.hpp>
 #include <hpx/schedulers/lockfree_queue_backends.hpp>
 #include <hpx/schedulers/queue_holder_numa.hpp>
 #include <hpx/schedulers/queue_holder_thread.hpp>
@@ -26,7 +27,6 @@
 #include <hpx/threading_base/thread_num_tss.hpp>
 #include <hpx/threading_base/thread_queue_init_parameters.hpp>
 #include <hpx/topology/topology.hpp>
-#include <hpx/util_fwd.hpp>
 
 #include <array>
 #include <cstddef>
