+++ libs/schedulers/include/hpx/schedulers/thread_queue_mc.hpp
@@ -14,9 +14,9 @@
 #include <hpx/datastructures/tuple.hpp>
 #include <hpx/errors.hpp>
 #include <hpx/functional/function.hpp>
-#include <hpx/runtime/threads/policies/deadlock_detection.hpp>
-#include <hpx/runtime/threads/policies/maintain_queue_wait_times.hpp>
+#include <hpx/schedulers/deadlock_detection.hpp>
 #include <hpx/schedulers/lockfree_queue_backends.hpp>
+#include <hpx/schedulers/maintain_queue_wait_times.hpp>
 #include <hpx/schedulers/queue_holder_thread.hpp>
 #include <hpx/schedulers/thread_queue.hpp>
 #include <hpx/thread_support/unlock_guard.hpp>
