+++ src/runtime/threads/threadmanager_counters.cpp
@@ -17,9 +17,9 @@
 #include <hpx/performance_counters/counters.hpp>
 #include <hpx/performance_counters/manage_counter_type.hpp>
 #include <hpx/runtime/thread_pool_helpers.hpp>
-#include <hpx/runtime/threads/policies/maintain_queue_wait_times.hpp>
 #include <hpx/runtime/threads/threadmanager.hpp>
 #include <hpx/runtime/threads/threadmanager_counters.hpp>
+#include <hpx/schedulers/maintain_queue_wait_times.hpp>
 
 #include <cstddef>
 #include <cstdint>
